requirements.txt
# Core dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0
jinja2>=3.1.2
python-multipart>=0.0.6

# LangChain and related
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.9
langchain_text_splitters>=0.0.1

# DashScope API (Alibaba Cloud)
dashscope>=1.10.0

# Vector database
faiss-cpu>=1.7.4

# Image processing
pillow>=10.0.0
pytesseract>=0.3.10

# PDF processing
pypdf>=3.15.1
pdfplumber>=0.10.2

# Excel handling
openpyxl>=3.1.2
pandas>=2.1.0

# Email functionality
# Note: smtplib is part of the standard library

# Web search
duckduckgo-search>=3.9.3

# Feishu integration
requests>=2.31.0

# Database
sqlalchemy>=2.0.20
# Note: sqlite3 is part of the standard library

# Utilities
tqdm>=4.66.1
# The following are part of the standard library and don't need to be installed:
# uuid, webbrowser, threading, hashlib, shutil, re 